Done with the optimizer. I will call it a day here.
Looking over the Leanring module, it is a mess right now. What I am going to do tomorrow is do some reorganization and then I am going to move to debugging. That is going to take me a while. No doubt I have errors out the wazoo, but most of them or even all should be of the type variety..

So far, the going has been slow, but these 800 lines did not come out of thin air, so I must be doing something right. I wish I could program 24/7 if only to get this done faster, but I need my rest.

I am definitely looking forwrd to seeing what the response to the tutorials will be. But no matter what, I doubt I will have the same uptake as Javascript, unless I produce something miraculous with the language.

Ultimately, Spiral is fast and expressive without equal, but that does not mean much by itself. It definitely is more complex to program in than F# for example. Given that, I need to make the blood flow for it to get recognition.

I am definitely glad now that instead of stopping at making it the Cuda backend for the ML library, I decided to go forward and make it a fully fledged language.

Without a doubt, my courage will be rewarded.