Today I woke up 1.5 hours later than usual, so I'll only be really starting now since I missed my usual morning routine. I've done a bit of reorganizing and the first thing I will do to start things off will be to put back the old tests. After that I will test the backwards pass.
I should be doing the bias kernels and the rest after that, but screw that. I had enough. After I am done with this test I will start work on the tutorials. It is weighting on my mind and I do not want to delay this anymore.

With that determination set, let get this out of the way.

After that the next phase will come. I will get the library out of the way and for once actually move forward towards my goal of making money and catching up to Deepmind. I am tired watching them have all the fun while I lay down the bricks.