I implement Lazy and redesigned Option.
I won't bother testing Lazy explicitly as I am decently sure it will work fine. It will get plenty of exercise during the Learning tests anyway.

When I am done with the backwards pass, I will definitely implement `.` as a binary operator.

Now, let me do these lazy numbers properly.

In the previous version in the library, I used lazy vars exclusively on the host side, but that was because all the types and the boilerplate for them was such a pain in the ass.

This time I can do it properly, so let me do it.

Once I am done with this, I shoudl be able to start testing the backwards pass.