"        /// Used for dividing by minibatch size.         inl scalar_div_by_outermost_dim cost_function x =             inm y = cost_function x             inm batch_size = to (type (primal y)) (primal x .dim |> fst)             scalar_div y batch_size
Now I can remove this for real. Useless crap. I never like it.

...Fuuuuuuuuu…

Yeah, I am going to stop here. I think I made up for getting up a bit later than usual today.
As it turned out, there were some typing issues with dual numbers and architecture problems related to that. I did not manage to do the backwards pass today, but I think I have it all set up today.

All in all, today was pretty wonderful and I was in the zone the whole time. I won't be long until I finish this with days like these. But nearly all days are like this aren't they? It makes one wonder why I am so afraid of diving into programming.

Somehow, when I go at it with the intention of programming, that makes it harder to get into zone.

To program effectively, one must focus on the smallest possible point of action and act on it. Small successes make it possible to get into the flow."