As soon as I unwound an idea came to me and I simply had to implement it. What I had previously in map_redo_map was retarded. I chopped off that map at the end and implement hostlazy_map and host_map.
Strangely enough the idea for it came to me before I even started this, but not the concrete plan, so I went with the dumb idea of adding another map to map_redo.

I guess I first had to rewrite it in order to draw out the duplication. Ahhhh...

Well, at least it looks nice now. I have nothing to be ashamed of anymore, though I may yet modify this so as to not depend on dead code removal of the F# and C++ compilers, but for now this will do fine.

Once again, I am really happy to have finally gotten that divide by batch size issue under control.

Tomorrow I have a really good chance of getting a significant chunk of this done if not all.

Let me finally call it a day.