Let me start the day. Am I missing something?
There is a bunch of stuff I need to do before I can declare the backwards pass to be complete, but I will leave them for later. I want to do the tutorials and document the language properly.

So with that in mind, let me try running it for the first time. I will redesign the map functions after I get it to converge on Mnist.