Right now I am getting a type error in the two branches of map_redo, so I am going to redo the function to accept a neutral element.
It seems that way of doing things is not good after all. I was going to run into trouble later with my current way of doing this when the time comes to do warp reduces on undersized columns.

I know what th error is too, it is that damn map function that is causing the problems in final_reduce. Nothing else makes sense.

I did not like the current design of map_redo anyway. In general I do like branches as they unsafe and hard to get right.

Well, with that said, let me go have breakfast. I'll continue after that.