I can't believe I finally got it to work. It worked fine the first time once i got through all the type errors.
This is a massive succeed. At long last, at long last I got the bakcwards pass to work!

There is so much work left to be done, but after trying and failing to master the basic design in the F# library, I finally got a working prototype in Spiral.

In terms of what I've done with the new language, I am at the point where I was exactly two years ago. It was in December of 2015 when after trying to do AD by hand for months, I finally discovered how to do it using closures and it made a significant impact on my development as a programmer.

It was for the first time I really realized the power of abstraction. Up to that point I've been using F# much like I would have a more convenient C++.

And in 2018 I discovered partial evaluation and learned how to harness its power. 2019 will not be like the past two years.

I have the two supremely powerful tools in AD and PE. And the Spiral language.

Tomorrow I will start on the tutorials. Even if it takes me two months to finish it, I don't care. It will be done.

Even if I do nothing else, Spiral will be my one contribution to this world. Simulacrum is my soul and Spiral is my power. Let might make right.