"Test99 passes and has been added to the pile.
With this let me commit and I will have a break. It is amazing that I could not do the parser yesterday due to fatigue.

But the failure did allow me to think it through during the night so it is not like it is a waste.

Today I will write the sub-chapter on join point recursion and hopefully post the first part of the tutorial. If not today then tomorrow.

There is quite a bit of material to do before I can cover the entirety of the language."