Ok, let me slowly start.
I got distracted by some of the replies on the Reddit threads I posted yesterday. There has not been much to be honest, which is roughly what I'd expect.

I find it ironic that after talking about Spiral's hygiene, I suddenly have a problem with that just as I'd released the language. Talk about bad luck. Well, it will even out over the long term.

During the night I really thought a lot about it, but as expected, hygiene breaking bugs are really on par with memory corruption bugs. I can't reason out where the error might be just from its effects.

I did think of a lot of strategies for narrowing it down though.

First I want to look carefully that the error is not in expr_prepass. Then I will slowly try to encircle the problem.

Also I had an idea of how to simplify PatClauses and redefine it in terms of PatAnd and PatOr + a new PatBody I will introduce.

The rest of my ideas are related to where to put in print statements.

I debugged this same problem on 11/6, 1.5 months ago and essentially resolved it by breaking type equality on functions. I need to fix this mistake.

I did a lot of debugging back then, but I could not figure out why the error happens, all I know that it is some strange breaking of hygiene.

Let me give a whack at it. Forget the tutorials and forget anything else for the time being.