Let me start working on the Loops chapter. I am going to do Loops, Arrays and Lists in that order. I can't skip over them, and into Parsing and Tensors.
Also for the past few days I've been using Dillinger to edit markdown. It turns out VS Code which I already had installed is much better at that so I am going to be using that instead.

I feel like doing it all at once, but I need to keep in mind that trying to do it like that will just lead to frustration. I will give the tutorial time to mature.

Let me slowly get into the groove. Today also I might have some obligations to take care of, so I won't be able to spend working as much on the tutorial as I'd like.

Also I will take this chance to redesign the Loop module as well.

...Actually, let me merge the Array and the Loop chapters. After all, I need to use the loops on something.