It is about time to start the action. The first thing on the agenda is to finish the breakable loop.
Then it will be time to cover the basic array functions.

I kind of want to cover the Save The Princess problem, but am a bit conflicted how to do that one before covering the Monads and Parsing.

I'll have no choice to keep it for last.

So let me just start. I need to get into the groove.