I will take a break here. I am done with the Loops subchapter.
What I need to do right now is put the new `for` function into the library and modify the examples to reflect that.

After that I will take it a bit more slowly with the Arrays subchapter.

The Lists chapter won't be hard either. I'll cover Arrays and Lists in a breezy fashion without necessary complications that the Loops chapter had.

I considered doing HackerRank problems for the tutorials, but they have the notable problem of needing to be parsed and all of that which would clutter up the generated code. And I still haven't covered parsing yet, so forget that plan.

Yeah, Arrays, Lists, Parsing, Tensors, Cuda.

I can cover all of this ground maybe by the end of the year and then I will have to spend more time documenting the compiler and the libraries more in depth.

I won't force myself to do too much. I'll do this as long as I feel it is necessary