Let me stop here for a bit. Spiral is not Lisp, so things like querying structures should have their own patterns even for things that are not often used as a matter of setting the right example.
In the earlier days of Spiral development I was pretty much always adding features and debugging them, so I had started thinking that I would not be doing that again so soon now that the language is starting to come out of the shadows.

Let me take a break from the tutorials to think about this for a bit and then I will implement it. It won't be hard.