I will leave type as is as type get and instead make join_type.
I definitely do not want to make type the keyword for the join point as join points should not be used to get the types. They could, but they are more expensive operations than type get and will fill up the type dictionary rapidly if used on trivialities. Type join points should be used for recursive type definitions which are not that common.

Forget what the language is and focus on what it could be.

I will start by removing stuff the language. This will also allow me to simplify the parser. Right now I've really taken it apart to allow the different parsing behavior in type statements.

Well, let me get on with it. The scope of changes that I have to make here will not be hard. It won't take me long.