Some more changes. I dropped the () requirement for `type` and now it acts like the two join points syntactically.
The reason why I had that in the first place is because it was colliding with the `type` statement previously.

I really will call it a day with this.

I should have done it like this from the start, but hopefully now the language should syntactically coherent.

There is nothing that really bothers me about it anymore.

The list module on the other hand...

Yeah, I am thinking that it won't be as easy as the Array module. I need to stop thinking it will be.

I need to redo it from the start. Also the `lw` function has an error in it. Pitiful.

Well, I'll be able to deal with it tomorrow.