As expected the compiler got broken by my changes. Right now test10 is giving me completely unexpected results and I Am not sure what is going on here.
Edit: Forgot to push this earlier, but I am still making no headway on figuring this out.

I am currently in the state of complete perplexity. How complicated. This will take a while until I get that vital flash of insight.

Are my assumptions here broken somehow?

I'll take a break for a while to think about this. I am not sure where to even begin debugging and I've been thinking about this for over an hour now.