"12pm. Ok, let me resume. I am still thinking about this. match dyn (ab .A, ab .A, ab .A) with | .A, _, _ -> 1 | _, .B, _ -> 2 | .A, .A, _ -> 3 | _ -> 4
Here is a slightly smaller test.

...Let me take out everything apart from box and dyn from the main library. Actually...let me take out everything.
In pat_tuple the arg is  pat_var_3_0
In pat_tuple the arg is  pat_var_3_0
In pat_tuple the arg is  pat_var_3_0
Time for prepass: 00:00:00.0167270
[lit inside the case  pat_var_3_0 is, [var (union {type_lit (A) | type_lit (B)}), var (union {type_lit (A) | type_lit (B)}), var (union {type_lit (A) | type_lit (B)})]]
v  pat_var_3_9 is 0...TyV (0, UnionT (set [LitT (LitString "A"); LitT (LitString "B")]))

What the dick? So it is the main arg in pat_tuple just I had thought it would be.

Let me print out the name and the typed_expr of everything in the case_ then.

Let me go me just figure this out and after that I will go do the chores and take a break.

This is taking an amazing amount of brainpower from me.
v  pat_var_3_0 is TyList
  [TyV (0, UnionT (set [LitT (LitString "A"); LitT (LitString "B")]));
   TyV (1, UnionT (set [LitT (LitString "A"); LitT (LitString "B")]));
   TyV (2, UnionT (set [LitT (LitString "A"); LitT (LitString "B")]))]

Ohhhhhhhhh...

I am such an idiot.

Yeah, what I did with case was useless for this problem.

12:35pm. Apart from what I did for the term function I basically wasted half a day on this. And I am still not done wasting it yet.

Yeah, it is not that the idea did not occur to me previously. It probably did, but I discarded it.

12:45pm. I am drowning in the sea of assumptions. I am giong to backtrack to the previous thing. I will keep the term function pattern changes, and then I am going to write some tests to see exactly how active patterns work in tandem with join points and casing.

I do not like this. But there is really no way apart from rewriting left to right in order for me to implement patterns.

I am trying to think whether I cna salvage parts of the idea that I had today. Is applying in case doing anything?"