Let me resume. You know, instead let me just read the paper on the second Futamura projection.
This thing today took the wind out of my sails.

I'll do a bit of the tutorial today.

Jesus, there sure will be a lot of exceptions for the union types.

Ahhhhhh...but I cannot do them right. It is impossible. I'd need a fundamentally different approach to compiling them and I can't think of one. This one works well enough as it is.