I can't believe the string_stream function had a bug in it. I thought it was a compiler bug for a second. At this rate it does seem like I will be redesigning the parsing library as well.
I am just kidding. There is no way that will happen.

Edit: Forgot to push the above.

Let me stop here for a bit. I need to think about this for a while.

Writing this stuff is not as easy as it seems at first glance. Unlike with lists and arrays, parser are more complicated and the library is in monadic form from the start.

I cannot start there.

It will be a lot of work, but I need to retrace the steps I had taken when I first made it. That means making it first in the applicative style and then do it in the monadic style.

Let me rest for a bit while I gather my power and then I will do the int parser.

The one currently in the library is poorly designed. What an annoyance.

I really did very few things right and that really shows while I am doing these tutorials. Trying to explain the language to other people is definitely having an effect of improving it so I should keep doing it.

I will also get rid of some things that would otherwise be in the library such as that parser module.

Given that this will be a pain, the goal for the tutorial will be to make a parser that parses a sequence of ints into an array in both the applicative and monadic fashion.

That should be enough to demonstrate all the essential elements of the design.

...Actually, let me make it parsing a multidimensional array separated by commas. Maybe like the one I did for Futhark that went nowhere?

Given that I do not have error handling, maybe something like that would be ideal to try out.

I will not be easy. I'll decide once I am done with parser for a sequence.