I will stop here for the day.
The style I was using so far was working for me for the previous chapters, but I get the sense that this attempt was a failure. I should put in the parser descriptions and the type annotations and then start at that point and simply show how to rewrite them in monadic style.

Then I should prove that the applicative and monadic styles compile down to the same code and then compare them to F#.

Not every chapter needs to be book length. I am going to try refocusing my effort tomorrow.

Every chapter deserves at least a few days effort on them. I won't be able to finish the tutorial by new year, that is for sure.