Let me take out the stuff from the tutorial. I do not want people to start reading that only to get confused.
Chapter 4 will just be a benchmark and a short demo. I won't waste time explaining parsers.

They very simple to me, but that is only because I understand them so well. It would be entirely different to people unfamiliar to them. They are a fairly advanced subject and a significant amount of research on language development in the early years of the field was on that subject.

I should not make a mistake of looking down on it. If the users are that interested in it they just look into Fparsec's documentation for more info.

I'll get through this is in a day or so and then I will finally get on with the tensors.

That is where the real fun starts. I've been looking forward to that for a while now.