I will call it a day here. I have the Tensor chapter planned out and will be able to make good progress tomorrow. The whole parsing thing was a bit difficult.
At this rate I might be able to finish the documentation in 2-3 weeks. That might put it at over 10k lines by the end of it.

Seems reasonable. Yeah, going at it until the end of January seems like it would be enough as a time investment into this.

A language without documentation is pitiful and having it should provide a pillar of support for other users. So far, it has not made a splash in particular. One guy likes it, the other called me arrogant and ignorant and I am hoping some of the lurkers are interested enough to have a look at it.

Well, either way, I did make it for myself so the popularity of it does not matter to me.

Tensorflow and PyTorch and the rest of the frameworks around bound to collapse under their own weight sooner or later and I want to have nothing to do with them.

Yesterday when I saw the erroneous 80x gain, for the first time, I really thought that this was it. And 20x over F# as a baseline is not that far. For a language as expressive as Spiral, this kind of speed is miraculous. If C does not die, it is pretty much the other people's fault for letting it come to that. This is not like before when it was only competing with slow as mollases kind of languages.

Without a doubt, Spiral will give me the edge I need to finish the ML library and venture into the great beyond after that. Even in its half finished state it's power is can be seen shinning underneath. It just needs to be digged out.

I guess the only difference now from before is that other people would be reading these commits, so I should probably stop treating them as a diary. Well, it has been fun.