I am done for today. I went from 3.9k to 4.35k lines in the tutorial. With this the size of documentation finally exceeds the size of the compiler.
I also renamed `caseable_is` to `caseable_box_is`. It should have been like that from the start, and it is my fault that it was not.

As planned, tomorrow I will take a sabatical to write a yearly review for the Simulacrum blog. I have a single datapoint for Spiral's performance so I have a something to brag about, and I have a 100 pages (169 kb) of documentation, so I do not have to make stuff up just to have something to show.

I'll let it simmer overnight. It it time to wind off.