I am finally done with the Tensor chapter. Really, I did not have much left to go there. I did most of it 2 days ago.
So now comes the Cuda chapter.

I am starting to reach the endgame with regards to the tutorials here.

The thing is, this stuff was only done very recently so it is immature, so I do not have much to go on. So far, only two kernels have ever been written in Spiral.

Let me take a break to think about it.

I'll give it a shot describing the allocator maybe?