"http://lambda-the-ultimate.org/node/4345#comment-75494
"Fexprs are more expressive than macros, I agree. But it seems to me that the greater expressiveness of fexprs (relative to macros + first-class functions) can only be utilized in ways that reveal the implementation details of one module to another - even at runtime, as you mention.

My goal is to express useful, high quality, maintainable programs. A high level of expressiveness for useless, error-prone, or unmaintainable programs does not contribute towards my goals. Hence, I don't acknowledge expressiveness as a goal, good, or end of its own. I believe we can find more balanced approaches that achieve expressiveness without sacrificing effective control, though I think such approaches may be more specialized (compared to the one-abstraction-fits-all fexpr)."

Actually, I do agree with this. Yeah, fexprs suck. I definitely do not want to expose that level of detail to the user. Just think of how it would be to expose the meta-enviroments (join point dictionaries and the language environment.) It would be a complete and a utter mess. I as the language author do not want that crap poluting the host language. Furthermore, there are large issues with optimizing such languages, unlike Spiral which is used for optimization.

Forget Lisp. The way Spiral achieves uniformity is different from the Lisps.

It actually has many different features that all mesh well together. It's approach of taking a pure dynamic language and adding partial evaluation and abstract interpretation over first class types is its brilliance.

I won't mention Fexprs in the context of Spiral ever again.

It took me a while to decide this."

The above is an excerpt from my journal

So today I ended up not working on the tutorial. Also now tomorrow's update is fully complete and proofread. I almost said typechecked.

Actually I had intended to split the day between proofreading and work on the tutorial, but I started studying the source for the verious Kernel derivatives and that sidetracked me.

Well, taking a break every now and then to work on side things is ok.

That won't happen again. For the next two weeks, my focus will be on finishing the user guide. I have a lot to cover.

I think two weeks might be enough even though it might go up to the end of January.

Also I am going to have to do extra and add comments for all the stuff in the standard library and do the reference as well.

All of that will take time. Well, I won't force myself. I'll try to keep an even pace.

After I am done with the documentation it will be time to resume work on the library. I'll put off the Cuda tutorial for about six months. I have no intention of rushing it. I'll forget about the unfinished tutorials until I have more to show and focus on the rest.