I am done for the day. I do not feel like it anymore. 400 lines are quite enough at any rate for one day.
I started relatively late today, but as it turns out the amount of that one does per day is not proportional to effort, bu proportional to the effort done in the zone.

I can do about 8h of work per day, but I think my limit for being in the zone is closer to 4h unfortunately.

I also had some time to digest the material yesterday, so that should get added to the tally.

I need to get the user guide out of my system before I can do anything with the language. Once I am done with documentation, I think I will shop it around for a bit and see what the other guys think about it.

My sense that the PL guys are more interested in their own languages than my own which is understandable.

I think Spiral is absolutely groundbreaking, but I want the ask the guys at Ltu what they think of it first.

I think I might post it on /r/programming as well on a couple of other subs.