500 per day is enough. I am out of energy.
I need to think of what is next. Tomorrow, I will cover the prepass and pattern compilation at least.

I'll also try squeeze in the layout types.

Though if I managed that, at that point wouldn't I have covered the core topics of the compiler?

Who knows, at this rate maybe two weeks will in fact be enough to finish the documentation? Well, good if that is the case. I really want to get back to work on the library.

Edit: Forgot to push this. I am done.