That is 170 lines added and with it I am done with the layout types and the macro chapter.
I am also done with the user guide.

Ahhh, it has been a long trip hasn't it?

It is amazing how once the curtain is raised and the tricks are revealed how trivial the whole thing is.

Did it really have to take me this long to figure it out? Soul crushing simplicity is trully the best part of science.

I'll probably have to shove the idea down people's throats.

I think for Spiral the ideal scenario is that it gets a few good contributors who recognize its worth and intend to use it for themselves. That way we could pool our power and use it to enrich ourselves.

Though I am showing it around, I am really not that interested in serving the wider community.

That would just be a chore and an unpaid burden for me. I'll do it if comes to that, but I do not want to make it seem as if my time is free.

I will answer any questions regarding Spiral as it would be good feedback and a way to improve the language.