Yesterday I ended on a short note as I'd rather eat than spend time composing a commit, but today I am ready for action.
Today I won't do any writing - I am done with the documentation. I am not under the illusion that the thing is in any way done, but is sufficiently done.

I'd prefer it if the tutorials were longer, but this will have to do for now as I am out of interesting material.

Instead what I will do today is go from top to bottom and try to spot any typos and grammar errors. Once that is done I will post a link to the language on the /r/machinelearning sub just to see if those guys can see the value in the language on its own merits.

I need this in order to decide in which direction I want the language to go.

If I don't manage to attract any contributors for the langauge, then I will just go with my original plan of keeping it as my personal tool.

This does have some advantages of not having to burden myself with maintainance and documentation, but I wanted to try it once. Whenever I see an excellent tool that is poorly documented I always think what a waste it is. I do not want Spiral to be overlooked because of my own neglect.