I am a bit worried that I may have chosen a wrong title for the post. Maybe "A language made for deep learning" would have been more catchy? I'll try that tomorrow if I get no views.
I am prepared for whatever outcome in terms of adoption, but it is no good if people on the sub skip it over completely.

Well, forget that.

It is time to finish proofreading the user guide. I'll do that today and finally resume work on the ML library tomorrow. Though I maybe take a short break.

Well, I always say that, but I never do it.

I definitely am sick of the documentation though. I wrote in the update that I would do it until the end of January, but so much for that.

Well, I did find a bug in the xor pattern and got to refactor a little so it was worth it.