Given that there is no difference on the tests now apart from test90 which just has a few tags swapped, I think I can assume that it works.
With this the redesign is done.

Beautiful.

Now I also need to redesign the CudaTensor, but it should be simpler now.

Simplicity gives real wins in the long run.

Well, let me take a look. It won't be too hard. After that I will start work on the library in earnest tomorrow.