I think it is time to slowly start. I am not done chilling yet, but let me write this first.
There is still that bug from yesterday lurking around, but for once I will just leave it be.

The real issue with the Learning module is that I have no tests. I had a few, but I somehow lost them. This just shows how clunky the Learning module is right now.

The reason why this happened is because memory allocation is forcing me to couple all the code explicitly, which makes testing it difficult.

The first order of bussiness will be to take care of that.

For the next few days, my focus will be on simply improving the organization of the library and putting in a testing harness so I do not have to worry about refactoring breaking stuff again.

Looking back, I was really poor at testing throughout the last 3 years, but I think I can do much better now thanks to all the experience I've accumulated.

I am not a beginner anymore, not after all this effort torwards improving myself.

https://www.reddit.com/r/ProgrammingLanguages/comments/7nsrn2/monthly_what_are_you_working_on_how_is_it_coming/ds9ng0f/?st=jc33eh2r&sh=a3f4efa7

Here is my new year's resolution.

I intend to get into it slowly, just as usual/