The first 3 tests pass. I will admit that it is really heavy going. I am finding it hard to stay motivated during this refactoring.
I've been through this before a significant amount of times and they way to get past this is always the same - work hard. If work cannot be done easily it will have to be done hardly.

It is not that this kind of work is particularly difficult on its own even though emulating the stack allocation of GPU memory is introducing all sort of couplings. It's just that the fire needs some coaxing to start.

Tough words in the resolution I posted are just bluster - this is the real deal.

I will get through this. After all, this is a matter to be resolved between me and the computer. I am not depending upon the others over which I have no control for feedback.

I have it right here.

Let me have some air and then I will continue this. I want to at least do the kernels today.