Let me stop here for the day.
Today I was steady, but I can't say I entered the zone. I spent most of the day trying to get unstuck in fact.

I did fix a lot of core design failures from the last time around.

Just check this out...

"5:40pm.
    inl primal = function {DR={primal}} | primal -> primal
    inl adjoint = function {DR={adjoint}} -> adjoint | _ -> .nil

    inl primals = toa_map primal
    inl adjoints = toa_map adjoint

    inl (>>!) a b ret = a <| inl a -> b a ret

    inl is_unit = function
        | () -> false
        | _ -> true

    inl rec filter_units = function
        | x :: x' ->
            match filter_units x with
            | () -> filter_units x'
            | x -> x :: filter_units x'
        | {} & x ->
            module_filter (inl k (!filter_units (!is_unit x)) -> x) x
            |> inl x -> if eq_type x {} then () else x
        | x -> x

    // What this does is selectively filter out the results of applying f
    // where the adjoints are missing (in other words constants.)
    inl filter_based_on_adjoints x adjoint =
        inl rec mark_units = function
            | x :: x', y :: y' -> mark_units (x,y) :: mark_units (x',y')
            | (), () -> ()
            | (), _ | _, () -> error_type "Tuple dimesions do not match."
            | {} & x, {} & y -> module_map (inl k x -> mark_units (x,y k)) x
            | _, .nil -> ()
            | x, _ -> x
        mark_units (x, adjoint) |> filter_units

    inl filter_unit_and_branch x ret =
        match filter_units x with
        | () -> ()
        | x -> ret x
Ok, yeah, now this whole segment makes 100 times more sense.  Previuosly I was really terrified of missing an argument because I knew I botched the thing and had no idea what the hell was it even supposed to be doing. Here it all makes sense. I am no longer using `()` for empty adjoints because that conflicts with missing argument error.

These are all beginner mistakes in Spiral and in it not something my F# experience could have prepared me for.

Now let me do the map and map_redo. I won't test them today. As simple as this stuff is, it drained me completely.

I mean if figuring out the nil thing was so easy, I would have done it last time.

5:45pm. | {} & x, {} & y -> module_map (inl k x -> mark_units (x,y k)) x

Let me change this thing. I want to adjoints to decide the layout, not bck.

map_template .Add bck {in=primal; out} adjoint

Ah, yes I also need to figure this out. Hmmm...
                inl {in out adjoint} ->
                    toa_map ((|>) {in out}) bck
                    |> toa_map2 (+) adjoint
Actually, that `<|` should be `|>`

5:55pm. Let me improve toa_map2 and toa_map3 just a bit.
inl toa_map3 f a b c =
    inl rec loop = function
        | x, y, z when caseable_box_is x || caseable_box_is y || caseable_box_is z -> f x y z
        | (), (), () -> ()
        | x :: xs, y :: ys, z :: zs -> loop (x,y,z) :: loop (xs,ys,zs)
        | (), _, _ | _, (), _ | _, _, () -> error_type "Tuple dimensions do not match."

Yeah, this will give me some peace of mind.

I should adjust the tutorial.

6pm. Done. I am really pleased with this. Let me also add map_redo.

6:05pm. Note that doing it like this really make a lot more sense that having that mode hack directly in map_template. That was bad.
            match filter_units x with
            | () -> filter_units x'

Oh, yes, I can't filter only `()`. I should also do the same for .nil so at to avoid having to call filter_based_on_adjoints on itself.

6:15pm.
        inl map_redo {fwd bck} in ret =
            inl primal, adjoint = primals in, adjoints in
            inb !dr_lazyhost out = map_redo fwd primal
            ret (out, inl _ ->
                inl out = toa_map2 (inl P A -> {P A = A ()}) out.primal.value out.adjoint
                inl bck =
                    inl adjoint = type adjoint
                    inl bck = filter_based_on_adjoints bck adjoint
                    inl {in adjoint} ->
                        toa_map ((|>) {in out}) bck
                        |> toa_map2 (+) adjoint
                inb adjoint = filter_unit_and_branch adjoint
                map bck {in=primal; adjoint} adjoint
                )
Yeah, this should do it. map_redo really has a similar structure to map, but I've learned my lesson with join points. Sometimes similar things cannot be abstracted.

When that it the case you know you are dealing with an integral part of reality."

I am an amateur in my own language it seems.

But as difficult and simple as the things I did today were, they are going to have a tremendous effect on the integrity of the whole library going forward.

Not to mention the tests I am doing.

This stuff is important

This is the kind of thing one would expect a professional to be doing.

...Fuuu.

Really the map and redo_map are such marvels of engineering.

Writing them would be absolutely impossible in any other language, but Spiral.

I wish I had this sort of power 3 years ago, but it better late than never.

Tomorrow my aim will be to deal with the primitives and then move to the layers the day after that.

One thing I want to do is factor the run function.

It is definitely way too coupled right now.

I need to separate it more.

Also I want to redo the load_mnist function so it makes more sense.

A straight translation to Spiral from F# is not the best match.

I know how to do it better and I'll make time for it when it comes to that.

For now let me take a well deserved rest. I won't conquer the world in a day.